+++ CHANGES.md
@@ -1,3 +1,16 @@
+### Version 565 - (2020-06-05)
+
+- Fix pushing tags.
+- Fix formatting.
+- Rename fs.fstype attribute to fs.type.
+- Add docstring.
+- Improve plugin docstrings.
+- Fix attribute fs.type in example.
+- Improve plugin docstrings and remove "fixme" plugin.
+- Improve docstring.
+- Improve docstring.
+- Show help as manpages.
+
 ### Version 564 - (2020-06-04)
 
 - file plugin: Add file.xattrs attribute.